# docs/swagger.yaml
info:
  contact: {}
paths:
...
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Get an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              created_at:
                type: string
              id:
                type: integer
              order_items:
                items:
                  properties:
                    product_id:
                      type: integer
                    quantity:
                      type: number
                    total_price:
                      type: number
                    unit_price:
                      type: number
                  type: object
                type: array
              payment_method:
                type: string
              shipping_address:
                type: string
              total_price:
                type: number
              user_id:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get an order by ID
      tags:
      - Order
  /order/all:
    get:
      consumes:
      - application/json
      description: Get all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              orders:
                items:
                  properties:
                    created_at:
                      type: string
                    id:
                      type: integer
                    order_items:
                      items:
                        properties:
                          product_id:
                            type: integer
                          quantity:
                            type: number
                          total_price:
                            type: number
                          unit_price:
                            type: number
                        type: object
                      type: array
                    payment_method:
                      type: string
                    shipping_address:
                      type: string
                    total_price:
                      type: number
                    user_id:
                      type: string
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get all orders
      tags:
      - Order
  /order/create:
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          properties:
            order_items:
              items:
                properties:
                  product_id:
                    type: integer
                  quantity:
                    type: number
                  total_price:
                    type: number
                  unit_price:
                    type: number
                type: object
              type: array
            payment_method:
              type: string
            shipping_address:
              type: string
            total_price:
              type: number
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              created_at:
                type: string
              id:
                type: integer
              order_items:
                items:
                  properties:
                    product_id:
                      type: integer
                    quantity:
                      type: number
                    total_price:
                      type: number
                    unit_price:
                      type: number
                  type: object
                type: array
              payment_method:
                type: string
              shipping_address:
                type: string
              total_price:
                type: number
              user_id:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new order
      tags:
      - Order
  /order/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete an order
      tags:
      - Order
  /order/update:
    put:
      consumes:
      - application/json
      description: Update an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          properties:
            order_items:
              items:
                properties:
                  product_id:
                    type: integer
                  quantity:
                    type: number
                  total_price:
                    type: number
                  unit_price:
                    type: number
                type: object
              type: array
            payment_method:
              type: string
            shipping_address:
              type: string
            total_price:
              type: number
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              created_at:
                type: string
              id:
                type: integer
              order_items:
                items:
                  properties:
                    product_id:
                      type: integer
                    quantity:
                      type: number
                    total_price:
                      type: number
                    unit_price:
                      type: number
                  type: object
                type: array
              payment_method:
                type: string
              shipping_address:
                type: string
              total_price:
                type: number
              user_id:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update an order
      tags:
      - Order
  /order/user/{id}:
    get:
      consumes:
      - application/json
      description: Get all orders by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              orders:
                items:
                  properties:
                    created_at:
                      type: string
                    id:
                      type: integer
                    order_items:
                      items:
                        properties:
                          product_id:
                            type: integer
                          quantity:
                            type: number
                          total_price:
                            type: number
                          unit_price:
                            type: number
                        type: object
                      type: array
                    payment_method:
                      type: string
                    shipping_address:
                      type: string
                    total_price:
                      type: number
                    user_id:
                      type: string
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get all orders by user ID
      tags:
      - Order
...
swagger: "2.0"